requirements.txt
PyYAML==6.0.1
requests==2.31.0
feedparser==6.0.11
Jinja2==3.1.3
duckdb==1.0.0
mcp>=1.0.0
tenacity==8.2.3
structlog>=24.1.0
pytest-cov>=4.0.0
